fix build break on x86_64, couple UI fixes
git-svn-id: http://svn.sixgill.org/trunk@63 535b6189-6128-4a1a-9bee-cfdcb02e0d70

+++ imlang/block.cpp
@@ -294,7 +294,7 @@ BlockCFG* BlockCFG::Read(Buffer *buf, bool clone)
     }
     case TAG_Version: {
       Try(res);
-      size_t version;
+      uint32_t version;
       Try(ReadTagUInt32(buf, TAG_Version, &version));
 
       res->SetVersion(version);
